Level editor, parallax redux(again), dk level.
Did a lot of work on the level editor. Made mouse drags scale with the
camera's scale. Redid parallax scrolling to move at 50% of the speed the
camera does.  Make sure to add padding around the outside of the
background image. Also began work on a copy of the first level from that
WiiU Donkey Kong game.